requirements/requirements-dev.txt
# package dependencies
Pillow==9.0.1
colormath==3.0.0